rename: exception 패키지 구조 개선
내부/외부 예외, enum, 공통 패키지로 구조 세분화
