+++ src/main/java/com/sparta/spartatigers/domain/auth/service/AuthService.java
@@ -2,11 +2,10 @@ package com.sparta.spartatigers.domain.auth.service;
 
 import com.sparta.spartatigers.domain.auth.model.Token;
 import com.sparta.spartatigers.domain.auth.model.TokenClaim;
-import com.sparta.spartatigers.domain.user.model.LoginUser;
 import com.sparta.spartatigers.domain.user.model.User;
 import com.sparta.spartatigers.domain.user.repository.UserRepository;
-import com.sparta.spartatigers.global.exception.ExceptionCode;
-import com.sparta.spartatigers.global.exception.InvalidRequestException;
+import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
+import com.sparta.spartatigers.global.exception.internal.InvalidRequestException;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
