+++ src/main/java/com/sparta/spartatigers/domain/item/model/ItemCategory.java
@@ -2,8 +2,8 @@ package com.sparta.spartatigers.domain.item.model;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 
-import com.sparta.spartatigers.global.exception.ExceptionCode;
-import com.sparta.spartatigers.global.exception.InvalidRequestException;
+import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
+import com.sparta.spartatigers.global.exception.internal.InvalidRequestException;
 
 import java.util.stream.Stream;
