+++ src/main/java/com/sparta/spartatigers/domain/item/repository/ItemRepository.java
@@ -2,8 +2,8 @@ package com.sparta.spartatigers.domain.item.repository;
 
 import com.sparta.spartatigers.domain.item.model.Item;
 import com.sparta.spartatigers.domain.item.model.ItemStatus;
-import com.sparta.spartatigers.global.exception.ExceptionCode;
-import com.sparta.spartatigers.global.exception.InvalidRequestException;
+import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
+import com.sparta.spartatigers.global.exception.internal.InvalidRequestException;
 
 import jakarta.persistence.LockModeType;
 import java.time.LocalDate;
