+++ src/main/java/com/sparta/spartatigers/domain/user/service/UserService.java
@@ -3,8 +3,8 @@ package com.sparta.spartatigers.domain.user.service;
 import com.sparta.spartatigers.domain.auth.service.PasswordEncoder;
 import com.sparta.spartatigers.domain.user.model.User;
 import com.sparta.spartatigers.domain.user.repository.UserRepository;
-import com.sparta.spartatigers.global.exception.ExceptionCode;
-import com.sparta.spartatigers.global.exception.InvalidRequestException;
+import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
+import com.sparta.spartatigers.global.exception.internal.InvalidRequestException;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
