+++ src/main/java/com/sparta/spartatigers/global/exception/internal/CustomException.java
@@ -1,5 +1,6 @@
-package com.sparta.spartatigers.global.exception;
+package com.sparta.spartatigers.global.exception.internal;
 
+import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
 import lombok.Getter;
 
 @Getter
