+++ src/main/java/com/sparta/spartatigers/global/exception/internal/InvalidRequestException.java
@@ -1,12 +1,13 @@
-package com.sparta.spartatigers.global.exception;
+package com.sparta.spartatigers.global.exception.internal;
 
+import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 
 @Getter
 @RequiredArgsConstructor
-public class ServerException extends BaseException {
+public class InvalidRequestException extends BaseException {
 
     private final ExceptionCode exceptionCode;
