+++ src/test/java/com/sparta/spartatigers/global/config/FCMServiceIntegrationTest.java
@@ -3,8 +3,8 @@ package com.sparta.spartatigers.global.config;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-import com.sparta.spartatigers.global.exception.ExceptionCode;
-import com.sparta.spartatigers.global.exception.FirebaseException;
+import com.sparta.spartatigers.global.exception.enums.ExceptionCode;
+import com.sparta.spartatigers.global.exception.external.FirebaseException;
 import com.sparta.spartatigers.global.firebase.FCMService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
